Finalizing the project.
I removed the Util code since it was not relevant to the project any longer. I added some comments for classes added later in the project. I added a simple Lox script in case anyone is interested in checking that it works. Basically just a bunch of last minute cleanup.